Wire filter callbacks & Display search status & Cancelling
* Wire filter callback to keeping filter info in session data state.
* Implement search status in search bar
* Introduce entity to keep track for long running process.
* Cancel search operation safely.
* Cancel running operations on session closing.
* TODOs and notes

+++ application/apps/indexer/gui/application/src/session/ui/logs_table/mod.rs
@@ -19,7 +19,7 @@ impl LogsTable {
 
         let table = egui_table::Table::new()
             .id_salt(id_salt)
-            .num_rows(data.logs_count as u64)
+            .num_rows(data.logs_count)
             .columns(Self::text_columns())
             .auto_size_mode(AutoSizeMode::Never)
             .num_sticky_cols(1);
